docs/agent-repositories/aos-mcp-servers/pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "aos-mcp-servers"
version = "4.0.0"
description = "Config-driven MCP server deployment Azure Function app for AOS"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"

dependencies = [
    "aos-kernel[azure]>=3.0.0",
    "azure-functions>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pylint>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
